Add ModuleProvider
