+++ compiler/src/module/use_path.rs
@@ -1,0 +1,56 @@
+use super::module::{Module, ModuleKind};
+
+#[derive(Debug)]
+pub struct UsePath {
+    parent_navigations: usize,
+    path: String,
+}
+impl UsePath {
+    const PARENT_NAVIGATION_CHAR: char = '.';
+
+    pub fn parse(mut path: &str) -> Result<Self, String> {
+        let parent_navigations = {
+            let mut navigations = 0;
+            while path.starts_with(UsePath::PARENT_NAVIGATION_CHAR) {
+                navigations += 1;
+                path = &path[UsePath::PARENT_NAVIGATION_CHAR.len_utf8()..];
+            }
+            match navigations {
+                0 => return Err("The target must start with at least one dot.".to_string()),
+                i => i - 1, // two dots means one parent navigation
+            }
+        };
+        let path = {
+            if !path.chars().all(|c| c.is_ascii_alphanumeric() || c == '.') {
+                return Err("the target name can only contain letters and dots".to_string());
+            }
+            path.to_string()
+        };
+        Ok(UsePath {
+            parent_navigations,
+            path,
+        })
+    }
+
+    pub fn resolve_relative_to(&self, current_module: Module) -> Result<Module, String> {
+        let kind = if self.path.contains('.') {
+            ModuleKind::Asset
+        } else {
+            ModuleKind::Code
+        };
+
+        let mut path = current_module.path;
+        for _ in 0..self.parent_navigations {
+            if path.pop().is_none() {
+                return Err("too many parent navigations".to_string());
+            }
+        }
+        path.push(self.path.to_string());
+
+        Ok(Module {
+            package: current_module.package,
+            path: path.clone(),
+            kind,
+        })
+    }
+}
